test(data-structures): add unit tests for ordered_event_log
